@@ -18,10 +18,7 @@ final class MethodAccessorTest extends TestCase
     public function testGetValue(): void
     {
         $object = new class() {
-            /**
-             * @var string
-             */
-            private $name;
+            private ?string $name = null;
 
             public function getName(): string
             {
@@ -44,10 +41,7 @@ final class MethodAccessorTest extends TestCase
     public function testHasValue(): void
     {
         $object = new class() {
-            /**
-             * @var string
-             */
-            private $name;
+            private ?string $name = null;
 
             public function hasName(): bool
             {
@@ -70,10 +64,7 @@ final class MethodAccessorTest extends TestCase
     public function testIsValue(): void
     {
         $object = new class() {
-            /**
-             * @var string
-             */
-            private $name;
+            private ?string $name = null;
 
             public function isName(): bool
             {
