@@ -123,10 +123,7 @@ final class ValidConstraintTest extends TestCase
     private function getObject()
     {
         return new class() {
-            /**
-             * @var string
-             */
-            private $name;
+            private ?string $name = null;
 
             public function getName(): string
             {
