@@ -50,9 +50,7 @@ final class LazyValidationMappingProviderTest extends TestCase
         $container
             ->expects(self::any())
             ->method('get')
-            ->willReturnCallback(function (string $id) use ($services) {
-                return $services[$id];
-            })
+            ->willReturnCallback(fn (string $id) => $services[$id])
         ;
 
         return $container;
